requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn==1.6.1
xgboost>=2.0
mlflow>=2.14
pyyaml>=6.0
joblib>=1.3
fastapi>=0.110
uvicorn[standard]>=0.29
evidently>=0.5.5
pydantic>=2.6
pytest>=8.0
pre-commit>=3.7
shap==0.44.1 
scipy>=1.10
